% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_dict.R
\name{location_dict}
\alias{location_dict}
\title{Dictionary of clean location names (PL/CZ)}
\usage{
location_dict()
}
\description{
Return a dictionary used to clean the location data in the
 reports for Poland and Czech. The (key, value) scheme uses:
 (clean name, vector of misspelled names).
}
