% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smooth_variants_count}
\alias{smooth_variants_count}
\title{Internal functions}
\usage{
smooth_variants_count(
  df,
  k,
  date_column = "date",
  variant_column = "variant",
  count_column = "count"
)
}
\description{
Internal functions
}
